Switch to API client for models
