+++ build.gradle
@@ -35,6 +35,7 @@ def ktorVersion = "1.3.2"
 
 dependencies {
     implementation "com.kotlindiscord.database:database:+"
+    implementation "com.kotlindiscord.api:client:+"
 
     implementation "io.ktor:ktor-client-apache:$ktorVersion"
 
@@ -55,7 +56,6 @@ dependencies {
     implementation 'io.github.microutils:kotlin-logging:1.7.9'
     implementation 'io.sentry:sentry-logback:1.7.30'
     implementation 'org.codehaus.groovy:groovy:3.0.4'  // For logback config
-
 }
 
 application {
