+++ src/main/kotlin/com/kotlindiscord/site/components/Routing.kt
@@ -14,7 +14,6 @@ import io.ktor.routing.delete
 import io.ktor.routing.get
 import io.ktor.routing.post
 import io.ktor.routing.routing
-import io.ktor.util.getOrFail
 import io.ktor.util.pipeline.PipelineContext
 
 private val API_KEY = System.getenv("API_KEY")
@@ -30,13 +29,13 @@ fun route(body: suspend PipelineContext<Unit, ApplicationCall>.(Unit) -> Unit):
 
 fun apiRoute(body: suspend PipelineContext<Unit, ApplicationCall>.(Unit) -> Any?):
         suspend PipelineContext<Unit, ApplicationCall>.(Unit) -> Unit = {
-    // TODO: Validate API authentication
+    // TODO: Richer API auth
     // BODY: Once we have some kind of API auth validation, we should action it here.
     // BODY: We should also pass the auth details along, so the route itself can check perms and such.
 
-    val key = call.parameters.getOrFail("api_key")
+    val key = call.request.headers.get("X-Api-Key")
 
-    if (key != API_KEY) {
+    if (key == null || key != API_KEY) {
         call.respond(HttpStatusCode.Forbidden)
     } else {
         val resp = body.invoke(this, subject)
@@ -69,13 +68,13 @@ fun installRouting(app: Application) {
         get("/api/infractions", apiInfractionsGet)
         post("/api/infractions", apiInfractionsPost)
 
-        get("/api/users", apiUsersGet)
-        get("/api/users/{id}", apiUsersGetSingle)
-        post("/api/users", apiUsersPost)
-
         delete("/api/roles/{id}", apiRolesDelete)
         get("/api/roles", apiRolesGet)
         post("/api/roles", apiRolesPost)
+
+        get("/api/users", apiUsersGet)
+        get("/api/users/{id}", apiUsersGetSingle)
+        post("/api/users", apiUsersPost)
 
         get("/docs", redirect("/"))
         get("/resources", redirect("/"))
